matrix_voice: Cleanup and modernize.
